revised setup page components
