+++ src/routes/Settings.svelte
@@ -13,7 +13,7 @@
 
 <div class="ml-20">
   <div class="flex flex-col h-5/6 p-8">
-    <Alert color="dark" rounded={false}>
+    <Alert color="dark" rounded={false} accent>
       <span class="text-lg font-medium text-black-900 dark:text-blue-800 mb-4"
         >Settings</span
       >
